requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pymysql
pandas
httpx
reportlab
pydantic